backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
alembic==1.13.1
pytest==7.4.3
httpx==0.27.0
bcrypt==4.1.2
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
email-validator==2.1.0 